auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[argon2]
python-jose[cryptography]
python-multipart
